requirements.txt
# Backend Solo Nuttro - Dependências
# Framework
fastapi==0.110.1
uvicorn==0.25.0
python-multipart==0.0.20
python-dotenv==1.2.1

# Database - PostgreSQL
sqlalchemy>=2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
bcrypt==4.1.3

# Google Auth
google-auth==2.43.0
google-auth-oauthlib==1.2.0

# LangChain + LangGraph - IA
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langgraph>=0.2.0
langgraph-sdk>=0.2.0

# OpenAI/OpenRouter
openai>=1.0.0

# Image Processing
pillow>=10.0.0

# HTTP Client
httpx>=0.27.0

# Real-time
python-socketio>=5.11.0

# Utils
pydantic>=2.0.0
pydantic-settings>=2.0.0
orjson>=3.9.0

# Evolution API (WhatsApp)
aiohttp>=3.9.0

